+++ classes/commoneer/controller/auth.php
@@ -21,30 +21,18 @@ class Commoneer_Controller_Auth extends Commoneer_Controller_Template
 	 */
 	public $dash = 'dash';
 
-
-	/**
-	 * The login view for unauthenticated users
-	 */
-	public function action_index()
-	{
-		// Don't allow already authenticated users access to this page
-		if (Auth::instance()->logged_in()) {
-			$this->request->redirect('');
-		}
-	}
-
 	/**
 	 * Shows the login page and handles login
 	 *
 	 * @return void
 	 */
-	public function action_login()
+	public function action_index()
 	{
 		// Can't authenticate twice
 		if (Auth::instance()->logged_in()) {
 			$this->request->redirect('');
 		}
-		
+
 		// If the login form was posted...
 		if ($this->request->post()) {
 
@@ -54,10 +42,10 @@ class Commoneer_Controller_Auth extends Commoneer_Controller_Template
 			} else {
 				Notify::msg('Authentication failed!', 'error');
 			}
-		}
 
-		// Auth failed
-		$this->request->redirect($this->_login_url);
+			// Auth failed
+			$this->request->redirect($this->_login_url);
+		}
 	}
 
 	/**