+++ config/app.php
@@ -1,0 +1,31 @@
+<?php defined('SYSPATH') or die('No direct access allowed.');
+/**
+ * App settings
+ *
+ * @since 1.4
+ * @package Commoneer
+ * @author Ando Roots 2012
+ */
+return array(
+
+	/**
+	 * The title of the application. Think HTML <title>
+	 */
+	'title' => 'My Application',
+
+	/**
+	 * The description of the application. Think HTML <meta>
+	 */
+	'description' => ', fast and usable mood tracker focused on the individual.',
+
+	/**
+	 * The codename of the application
+	 */
+	'codename' => 'Mooditude',
+
+	/**
+	 * Current app version
+	 */
+	'version' => '1.0.0',
+
+);